+++ src/schema/resolvers/conference/conferenceResolvers.js
@@ -40,6 +40,10 @@ export default {
       const news = await News.query().where('conference_id', id);
       return news;
     },
+    activities: async ({ id }, data, { models: { Activity } }) => {
+      const activities = await Activity.query().where('conference_id', id);
+      return activities;
+    },
   },
   Query: {
     getAllConferences: async (
@@ -244,38 +248,18 @@ export default {
     deleteConference: async (
       root,
       { id },
-      {
-        models: { Conference, ConferenceTopic, ConferenceAttendee, News },
-        ValidationError,
-      },
+      { models: { Conference }, ValidationError },
     ) => {
       try {
-        // delete ConferenceTopic with conference_id
-        const confTopicWithConfId = Conference.query().where('address_id', id);
-        if (confTopicWithConfId) {
-          await ConferenceTopic.query()
-            .delete()
-            .where('conference_id', id);
-        }
-        // delete ConferenceAttendee with conference_id
-        const confAttenWithConfId = ConferenceAttendee.query().where(
-          'conference_id',
-          id,
-        );
-        if (confAttenWithConfId) {
-          await ConferenceAttendee.query()
-            .delete()
-            .where('conference_id', id);
-        }
-        // delete News with conference_id
-        const newsWithConfId = News.query().where('conference_id', id);
-        if (newsWithConfId) {
-          await News.query()
-            .delete()
-            .where('conference_id', id);
-        }
-
         const conference = await Conference.query().findById(id);
+
+        // // delete ConferenceTopic with conference_id
+        // // delete ConferenceAttendee with conference_id
+        // // delete News with conference_id
+        // // delete Activity with conference_id
+        await conference.deleteAllRelationship();
+        if (!conference) throw new ValidationError('Not found conference');
+
         await Conference.query().deleteById(id);
         return conference;
       } catch (e) {
